requirements.txt
pytesseract
pillow
tkinterdnd2